backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Transcription services
openai-whisper==20231117
openai==1.12.0
google-cloud-speech==2.24.0

# LLM services
google-generativeai==0.4.0
httpx==0.26.0

# Google Drive
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Audio/Video processing
ffmpeg-python==0.2.0

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0

# Background tasks
celery==5.3.6
redis==5.0.1
